added abs for int; changed float64 to int for square struct; reorganized structs & moves to separate file moves.go; renamed moves_test.go; renamed controllers.go
