+++ validation/controllers.go
@@ -4,7 +4,6 @@ package validation
 
 import (
 	"fmt"
-	"math"
 	"net/http"
 	"strconv"
 
@@ -101,6 +100,24 @@ func Simple(w http.ResponseWriter, r *http.Request) {
 	w.WriteHeader(http.StatusBadRequest)
 }
 
+// Advanced сервис отвечает за сложную валидацию хода по начальной и конечной клетке, а также по текущему состоянию
+// доски в нотации FEN. Также принимает на вход URL-параметр newpiece (это новая фигура, в которую нужно превратить
+// пешку при достижении последнего ряда, в формате pieceВозвращает заголовок HttpResponse 200 (ход валиден) или
+// HttpsResponse 403 (ход невалиден). Возвращает в теле JSON с конечной доской board в форате FEN.
+func Advanced(w http.ResponseWriter, r *http.Request) {
+	// TODO написать логику
+}
+
+// AvailableMoves сервис отвечает за оплучение всех возможных ходов для данной позиции доски в нотации FEN и начальной клетке.
+// Возвращает заголовок HttpResponse 200 (в случае непустого массива клеток) или HttpsResponse 403 (клетка пустая или
+// с фигурой, которой не принадлежит ход или массив клеток пуст). Возвращает в теле
+// JSON массив всех клеток, движение на которые валидно для данной фигуры.
+func AvailableMoves(w http.ResponseWriter, r *http.Request) {
+	// TODO написать логику
+}
+
+// Вспомогательные функции
+
 // parsePieceFromLetter переводит строковое представление фигуры типа k/q/r/b/n/p/K/Q/R/B/N/P в тип board.Piece.  Если
 // преобразование невозможно, возвращает ошибку.
 func parsePieceFromLetter(piece string) (board.Piece, error) {
@@ -133,196 +150,6 @@ func parsePieceFromLetter(piece string) (board.Piece, error) {
 		return 0, fmt.Errorf("%w", errPieceNotExist)
 	}
 }
-
-// Структуры данных
-
-// square клетка доски, моделирует ряд row и колонку column на шахматной доске в форматах float64
-type square struct {
-	row    float64
-	column float64
-}
-
-// newSquare создает новый экземпляр клетки доски square из представления s int8 пакета board
-func newSquare(s int8) square {
-	return square{
-		row:    float64(s / 8),
-		column: float64(s % 8),
-	}
-}
-
-// diffRow дает разницу в рядах между клетками s1 и s2
-func (s1 *square) diffRow(s2 square) float64 {
-	return s1.row - s2.row
-}
-
-// diffColumn дает разницу в колонках между клетками s1 и s2
-func (s1 *square) diffColumn(s2 square) float64 {
-	return s1.column - s2.column
-}
-
-// методы move
-
-// move для каждого типа фигуры. В случае невозможности сделать ход, возвращают ошибку, иначе возвращают nil.
-func move(piece board.Piece, from, to square) (err error) {
-	switch piece {
-	case board.WhitePawn, board.BlackPawn:
-		err = movePawn(piece, from, to)
-
-	case board.WhiteKnight, board.BlackKnight:
-		err = moveKnight(from, to)
-
-	case board.WhiteBishop, board.BlackBishop:
-		err = moveBishop(from, to)
-
-	case board.WhiteRook, board.BlackRook:
-		err = moveRook(from, to)
-
-	case board.WhiteQueen, board.BlackQueen:
-		err = moveQueen(from, to)
-
-	case board.WhiteKing, board.BlackKing:
-		err = moveKing(piece, from, to)
-	}
-
-	return err
-}
-
-// movePawn логика движения пешки. Может двигаться вверх (белый) или вниз (черный) на 1 или 2 клетки. Может съедать
-// по диагонали проходные пешки или фигуры соперника. Возвращает ошибку, если движение невалидно.
-func movePawn(piece board.Piece, from, to square) error {
-	var (
-		isVerticalValid bool // разрешено ли движение пешки по вертикали
-		isDiagonalValid bool // разрешено ли движение пешки по диагонали
-	)
-
-	switch piece {
-	case board.WhitePawn:
-		isVerticalValid = (from.diffColumn(to) == 0) && // верикаль не изменяется
-			(from.row != 0) && // пешка не может стартовать с 0 ряда
-			(to.row > 1) && // пешка не может прийти на 0 или 1 ряд
-			((from.diffRow(to) == -1) || (to.row == 3 && from.row == 1)) // движение вверх на 1 клетку, либо на 2 клетки (с 1 на 3 ряд)
-		isDiagonalValid = (from.row != 0) && // пешка не может стартовать с 0 ряда
-			(to.row > 1) && // пешка не может прийти на 0 или 1 ряд
-			(from.diffRow(to) == -1) && (from.diffColumn(to) == -1) // движение вверх по диагонали на 1 клетку
-
-	case board.BlackPawn:
-		isVerticalValid = (from.diffColumn(to) == 0) && // верикаль не изменяется
-			(from.row != 7) && // пешка не может стартовать с 7 ряда
-			(to.row < 6) && // пешка не может прийти на 7 или 6 ряд
-			((from.diffRow(to) == 1) || (to.row == 4 && from.row == 6)) // движение вниз на 1 клетку, либо на 2 клетки (с 6 на 4 ряд)
-		isDiagonalValid = (from.row != 7) && // пешка не может стартовать с 7 ряда
-			(to.row < 6) && // пешка не может прийти на 7 или 6 ряд
-			(from.diffRow(to) == 1) && (from.diffColumn(to) == 1) // движение вниз по диагонали на 1 клетку
-	}
-
-	if !isVerticalValid && !isDiagonalValid {
-		return fmt.Errorf("%w", errPawnMoveNotValid)
-	}
-	return nil
-}
-
-// moveKnight логика движения коня без привязки к позиции на доске. Может двигаться буквой Г. То есть +/- 2 клетки
-// в одном направлении и +/- 1 клетка в перпендикулярном направлении. Возвращает ошибку, если движение невалидно.
-func moveKnight(from, to square) error {
-	// разрешено ли движение конем на +/- 2 клетки в одном направлении и +/- 1 клетку в перпендикулярном ему направлении
-	isValid := (math.Abs(from.diffRow(to)) == 2 && math.Abs(from.diffColumn(to)) == 1) ||
-		(math.Abs(from.diffRow(to)) == 1 && math.Abs(from.diffColumn(to)) == 2)
-
-	if !isValid {
-		return fmt.Errorf("%w", errKnightMoveNotValid)
-	}
-	return nil
-}
-
-// moveBishop логика движения слона без привязки к позиции на доске. Может двигаться по всем диагоналям. Возвращает
-// ошибку, если движение невалидно.
-func moveBishop(from, to square) error {
-	// разрешено ли движение слоном по диагоналям
-	isValid := math.Abs(from.diffRow(to)) == math.Abs(from.diffColumn(to))
-
-	if !isValid {
-		return fmt.Errorf("%w", errBishopMoveNotValid)
-	}
-	return nil
-}
-
-// moveRook логика движения ладьи. Может двигаться вверх, вниз, влево, вправо на любое кол-во клеток. Возвращает
-// ошибку, если движение невалидно.
-func moveRook(from, to square) error {
-	// разрешено ли движение ладьей
-	isValid := (from.diffRow(to) == 0) || // по горизонталям
-		(from.diffColumn(to) == 0) // по вертикалям
-
-	if !isValid {
-		return fmt.Errorf("%w", errRookMoveNotValid)
-	}
-	return nil
-}
-
-// moveQueen для ферзя. Может двигаться вверх, вниз, влево, вправо на любое кол-во клеток. Может двигаться диагонально
-// на любое кол-во клеток. Возвращает ошибку, если движение невалидно.
-func moveQueen(from, to square) error {
-	errBishop := moveBishop(from, to) // может ли ферзь двигаться как слон по диагонаям
-	errRook := moveRook(from, to)     // может ли ферзь двигаться как ладья по вертикалям и горизонталям
-
-	if errBishop != nil || errRook != nil {
-		return fmt.Errorf("%w", errQueenMoveNotValid)
-	}
-	return nil
-}
-
-// moveKing логика движения короля. Может двигаться вертикально, горизонтально и диагонально только на одну клетку.
-// Также король из своего начального положения на доске (row 0 && column 4 для белого; row 7 && column 4 для черного)
-// может двигаться: на 2 клетки вправо или 2 клетки влево для рокировок.
-func moveKing(piece board.Piece, from, to square) error {
-	var (
-		isHorizontalValid bool // разрешено ли движение короля по горизонтали
-		isVerticalValid   bool // разрешено ли движение короля по вертикали
-		isDiagonalValid   bool // разрешено ли движение короля по диагонали
-		isCastlingValid   bool // разрешена ли рокировка
-	)
-
-	isHorizontalValid = (from.diffRow(to) == 0) && (math.Abs(from.diffColumn(to)) == 1)         // на 1 клетку вправо или влево
-	isVerticalValid = (math.Abs(from.diffRow(to)) == 1) && (from.diffColumn(to) == 0)           // на 1 клетку вверх или вниз
-	isDiagonalValid = (math.Abs(from.diffRow(to)) == 1) && (math.Abs(from.diffColumn(to)) == 1) // на 1 клетку по любой диагонали
-
-	// определение возможности рокировки
-	switch piece {
-	case board.WhiteKing:
-		if from.row == 0 && from.column == 4 && to.row == 0 && (math.Abs(from.diffColumn(to)) == 2) {
-			isCastlingValid = true
-		}
-	case board.BlackKing:
-		if from.row == 7 && from.column == 4 && to.row == 7 && (math.Abs(from.diffColumn(to)) == 2) {
-			isCastlingValid = true
-		}
-	}
-
-	if !isHorizontalValid && !isVerticalValid && !isDiagonalValid && !isCastlingValid {
-		return fmt.Errorf("%w", errKingMoveNotValid)
-	}
-	return nil
-}
-
-// http хендлеры
-
-// Advanced сервис отвечает за сложную валидацию хода по начальной и конечной клетке, а также по текущему состоянию
-// доски в нотации FEN. Также принимает на вход URL-параметр newpiece (это новая фигура, в которую нужно превратить
-// пешку при достижении последнего ряда, в формате pieceВозвращает заголовок HttpResponse 200 (ход валиден) или
-// HttpsResponse 403 (ход невалиден). Возвращает в теле JSON с конечной доской board в форате FEN.
-func Advanced(w http.ResponseWriter, r *http.Request) {
-	// TODO написать логику
-}
-
-// AvailableMoves сервис отвечает за оплучение всех возможных ходов для данной позиции доски в нотации FEN и начальной клетке.
-// Возвращает заголовок HttpResponse 200 (в случае непустого массива клеток) или HttpsResponse 403 (клетка пустая или
-// с фигурой, которой не принадлежит ход или массив клеток пуст). Возвращает в теле
-// JSON массив всех клеток, движение на которые валидно для данной фигуры.
-func AvailableMoves(w http.ResponseWriter, r *http.Request) {
-	// TODO написать логику
-}
-
-// Вспомогательные функции
 
 // ValidateMove обрабывает общую логику валидации хода
 func ValidateMove(b board.Board, from, to int) error {
